Add ability to recover volume from .phonons
But note that this is not useful for current
calcs as I'm using the cell file to change the
atom names. This changes the cell volume as seen
by the phonons utility (which is also why I have to
hack the atomic positions checker). The correnct
approach is to dump out an optimised .cell file
in the geometry optimisation step use this for the
phonons calc. There is a castep keyword to do this
which should be used. But for now, I'll need to
pull the volume from the .castep file.