Put alternitive x axis in place
Seems there is some kind of issue if we show values < 0 on the first
axis while doing the ugly work around having a discontinity in the
transform at values == 0. Avoid this by setting min value of 1/T^2
to 0 (which we map to 1E-10). This appears to work, at least, the
following seem to be in the right place!

  300 K -> 11.111
 2000 K -> 0.25
 1000 K -> 1.0
  400 K -> 4.0

NB: fiddling with these graphs is likly to get things wrong again!